requirements.txt
google-generativeai==0.3.2
supabase==2.3.0
Pillow>=9.0.0
requests==2.31.0